+++ common/config/chaintype.go
@@ -13,7 +13,7 @@ const (
 	ChainArbitrum        ChainType = "arbitrum"
 	ChainMetis           ChainType = "metis"
 	ChainOptimismBedrock ChainType = "optimismBedrock"
-	ChainXDai            ChainType = "xdai"
+	ChainGnosis          ChainType = "gnosis"
 	ChainCelo            ChainType = "celo"
 	ChainWeMix           ChainType = "wemix"
 	ChainKroma           ChainType = "kroma"
@@ -22,13 +22,13 @@ const (
 )
 
 var ErrInvalidChainType = fmt.Errorf("must be one of %s or omitted", strings.Join([]string{
-	string(ChainArbitrum), string(ChainMetis), string(ChainXDai), string(ChainOptimismBedrock), string(ChainCelo),
+	string(ChainArbitrum), string(ChainMetis), string(ChainGnosis), string(ChainOptimismBedrock), string(ChainCelo),
 	string(ChainKroma), string(ChainWeMix), string(ChainZkSync), string(ChainScroll)}, ", "))
 
 // IsValid returns true if the ChainType value is known or empty.
 func (c ChainType) IsValid() bool {
 	switch c {
-	case "", ChainArbitrum, ChainMetis, ChainOptimismBedrock, ChainXDai, ChainCelo, ChainKroma, ChainWeMix, ChainZkSync, ChainScroll:
+	case "", ChainArbitrum, ChainMetis, ChainOptimismBedrock, ChainGnosis, ChainCelo, ChainKroma, ChainWeMix, ChainZkSync, ChainScroll:
 		return true
 	}
 	return false
@@ -41,9 +41,6 @@ func (c ChainType) IsL2() bool {
 	switch c {
 	case ChainArbitrum, ChainMetis:
 		return true
-
-	case ChainXDai:
-		fallthrough
 	default:
 		return false
 	}